Refactor testsets, fixes #25
