+++ test/testsets/error_configuration/01.jl
@@ -1,4 +1,4 @@
-using Logging, Test, QuartoNotebookRunner
+include("../../utilities/prelude.jl")
 
 @testset "error_configuration/01" begin
     server = Server()
